% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasaurus-package.R
\docType{data}
\name{box_plots_long}
\alias{box_plots_long}
\title{Box plot data}
\format{
A data frame with 12420 rows and 2 variables:
\itemize{
\item \strong{Plot}: either the left, lines, normal, right or split boxplot
\item \strong{Values}: the corresponding values from each dataset
}
}
\usage{
box_plots_long
}
\description{
This dataset is the box plot data produced by Matjeka &
Fitzmaurice to demonstrate applicability of their process.
}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Same Stats, Different Graphs: Generating Datasets with
Varied Appearance and Identical Statistics through Simulated
Annealing. \emph{CHI 2017 Conference proceedings: ACM SIGCHI
Conference on Human Factors in Computing Systems.}
Retrieved from \url{https://www.autodeskresearch.com/publications/samestats}. #nolint
}
\keyword{datasets}
