% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a dataset}
\usage{
create_dataset(df, target_df, iters = 100, decimals = 2, shake = 0.2,
  max_temp = 0.4, min_temp = 0, allowed_dist = 2)
}
\arguments{
\item{df}{initial dataset}

\item{target_df}{target dataset}

\item{iters}{number of iterations}

\item{decimals}{allowed difference to the initial statistics}

\item{shake}{maximum amount of movement in each iteration}

\item{max_temp}{temperature at the beginning of the simulation}

\item{min_temp}{temperature at the end}

\item{allowed_dist}{allowed distance of new position to target dataset}
}
\value{
new dataset closer to the target one,
but with similar statistics compared to the initial df
}
\description{
Create a dataset with similar statistics to the initial dataset
and a shape close to a target dataset.
}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Same Stats, Different Graphs: Generating Datasets with
Varied Appearance and Identical Statistics through Simulated
Annealing. \emph{CHI 2017 Conference proceedings: ACM SIGCHI
Conference on Human Factors in Computing Systems.}
Retrieved from \url{https://www.autodeskresearch.com/publications/samestats}.
}
