% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{package}
\name{datasauRus}
\alias{datasauRus}
\alias{datasauRus-package}
\title{datasauRus}
\description{
datasauRus
}
\examples{
if(require(ggplot2)) {
  library(ggplot2)
  library(datasauRus)
  ggplot(datasaurus_dozen, aes(x = x, y = y, colour = dataset)) +
    geom_point() +
    theme_void() +
    theme(legend.position = "none") +
    facet_wrap( ~ dataset, ncol = 3)
}
}
