% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{perturb}
\alias{perturb}
\title{Function doing one round of perturbation}
\usage{
perturb(df, target_df, shake, allowed_dist, temp, x_bounds = c(0, 100),
  y_bounds = c(0, 100))
}
\arguments{
\item{df}{current dataset}

\item{target_df}{target df}

\item{shake}{maximum amount of movement in each iteration}

\item{allowed_dist}{allowed distance}

\item{temp}{current temperature}

\item{x_bounds}{bounds on x}

\item{y_bounds}{bounds on y}
}
\value{
new dataset after one round of perturbation
}
\description{
Function doing one round of perturbation
}
