% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{simpsons_paradox}
\alias{simpsons_paradox}
\title{Simpsons Paradox data}
\format{An object of class \code{data.frame} with 445 rows and 3 columns.}
\usage{
simpsons_paradox
}
\description{
Simpsons Paradox data
}
\keyword{datasets}
