% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{box_plots}
\alias{box_plots}
\title{Box plot data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2484 rows and 6 columns.}
\usage{
data(box_plots)
}
\description{
This dataset is the box plot data produced by Matjeka &
Fitzmaurice to demonstrate applicability of their process.
}
\examples{
summary(box_plots)

if(require(ggplot2)){
  ggplot(box_plots, aes(x=left))+
    geom_density()
  ggplot(box_plots, aes(x=lines))+
    geom_density()
  ggplot(box_plots, aes(x=normal))+
    geom_density()
  ggplot(box_plots, aes(x=right))+
    geom_density()
  ggplot(box_plots, aes(x=split))+
    geom_density()
}
}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Same Stats, Different Graphs: Generating Datasets with
Varied Appearance and Identical Statistics through Simulated
Annealing. \emph{CHI 2017 Conference proceedings: ACM SIGCHI
Conference on Human Factors in Computing Systems.}
Retrieved from \url{https://www.autodeskresearch.com/publications/samestats}.
}
\keyword{datasets}
