% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{box_plots}
\alias{box_plots}
\title{Box Plots data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2484 rows and 6 columns.}
\usage{
data(box_plots)
}
\description{
Box Plots data
}
\references{
Justin Matejka, George Fitzmaurice. (2017) Same Stats,
Different Graphs: Generating Datasets with Varied Appearance and
Identical Statistics through Simulated Annealing.
CHI 2017 Conference proceedings: ACM SIGCHI Conference on Human
Factors in Computing Systems.
}
\keyword{datasets}
