% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{box_plots}
\alias{box_plots}
\title{Box Plots data}
\format{An object of class \code{data.frame} with 2484 rows and 5 columns.}
\usage{
box_plots
}
\description{
Box Plots data
}
\keyword{datasets}
