% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{box_plots}
\alias{box_plots}
\title{Box Plots data}
\format{A data frame with 2484 rows and 6 variables:
\itemize{
\item \strong{left}: data pulled to the left
\item \strong{lines}: data with arbitrary spikes along a range
\item \strong{normal}: normally distributed data
\item \strong{right}: data pulled to the right
\item \strong{split}: split data
}}
\usage{
box_plots
}
\description{
A dataset demonstrating the utility of visualization. There are five different datasets
that all create the same boxplots.
}
\examples{
#save current settings
state <- par("mar", "mfrow")

par(mfrow = c(5, 2), mar=c(1,2,2,1))

nms <- names(box_plots)

for (i in 1:5){
  nm <- nms[i]
  hist(box_plots[[nms[i]]],
     breaks = 100,
     main = nm)
  boxplot(box_plots[[nms[i]]],
          horizontal = TRUE)
}


#reset settings
par(state)

}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Same Stats, Different Graphs: Generating Datasets with
Varied Appearance and Identical Statistics through Simulated
Annealing. \emph{CHI 2017 Conference proceedings: ACM SIGCHI
Conference on Human Factors in Computing Systems.}
Retrieved from \url{https://www.autodeskresearch.com/publications/samestats}.
}
\keyword{datasets}
