% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{datasaurus_dozen}
\alias{datasaurus_dozen}
\title{Datasaurus Dozen data}
\format{An object of class \code{data.frame} with 1847 rows and 3 columns.}
\usage{
datasaurus_dozen
}
\description{
Datasaurus Dozen data
}
\keyword{datasets}
