% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{datasaurus_dozen_wide}
\alias{datasaurus_dozen_wide}
\title{Datasaurus Dozen (wide) data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 142 rows and 26 columns.}
\usage{
datasaurus_dozen_wide
}
\description{
Datasaurus Dozen (wide) data
}
\keyword{datasets}
