% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{simpsons_paradox}
\alias{simpsons_paradox}
\title{Simpsons Paradox data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 444 rows and 3 columns.}
\usage{
data(simpsons_paradox)
}
\description{
Simpsons Paradox data
}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Same Stats, Different Graphs: Generating Datasets with
Varied Appearance and Identical Statistics through Simulated
Annealing. \emph{CHI 2017 Conference proceedings: ACM SIGCHI
Conference on Human Factors in Computing Systems.}
Retrieved from \url{https://www.autodeskresearch.com/publications/samestats}.
}
\keyword{datasets}
