% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{simpsons_paradox_wide}
\alias{simpsons_paradox_wide}
\title{Simpsons Paradox (wide) data}
\format{An object of class \code{data.frame} with 224 rows and 4 columns.}
\usage{
simpsons_paradox_wide
}
\description{
Simpsons Paradox (wide) data
}
\keyword{datasets}
