% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{twelve_from_slant_alternate_long}
\alias{twelve_from_slant_alternate_long}
\title{Twelve From Slant Alternate (long) data}
\format{An object of class \code{data.frame} with 2185 rows and 3 columns.}
\usage{
twelve_from_slant_alternate_long
}
\description{
Twelve From Slant Alternate (long) data
}
\keyword{datasets}
