% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{twelve_from_slant_alternate_wide}
\alias{twelve_from_slant_alternate_wide}
\title{Twelve From Slant Alternate (wide) data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 182 rows and 24 columns.}
\usage{
data(twelve_from_slant_alternate_wide)
}
\description{
Twelve From Slant Alternate (wide) data
}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Same Stats, Different Graphs: Generating Datasets with
Varied Appearance and Identical Statistics through Simulated
Annealing. \emph{CHI 2017 Conference proceedings: ACM SIGCHI
Conference on Human Factors in Computing Systems.}
Retrieved from \url{https://www.autodeskresearch.com/publications/samestats}.
}
\keyword{datasets}
