% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasaurus-package.r
\docType{data}
\name{twelve_from_slant_wide}
\alias{twelve_from_slant_wide}
\title{Twelve From Slant (wide) data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 182 rows and 24 columns.}
\usage{
twelve_from_slant_wide
}
\description{
Twelve From Slant (wide) data
}
\keyword{datasets}
